requirements.txt
google-api-python-client==2.100.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
html-to-markdown==1.3.2
beautifulsoup4>=4.13.4
tqdm==4.66.2
PyYAML==6.0.1